Add  abstract class for graphql operations
